requirements.txt
# VTOL Flight Simulation Visualization System
# Python dependencies

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
plotly>=5.0.0

# Configuration
pyyaml>=6.0

# Image and animation
pillow>=9.0.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# CLI and utilities
click>=8.0.0
tqdm>=4.62.0
